src/mcp/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
mcp>=0.1.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
pandas>=1.3.0
uvicorn>=0.15.0
fastapi>=0.68.0
yfinance>=0.2.3
